Allow override install options
It's conceivable that people may want to fork and utilize their own copy
of the aviary repository.  Depending on the environment where it's
being installed, various paths and other install-time options may need
to be adjusted.  Allow for this by accepting non-default values for
key installer options.

Also, disable existing cron entries with indication of what and when.

Signed-off-by: Chris Evich <cevich@redhat.com>